add function chart.depth.R
initial commit of function to chart market depth at a timestamp.  still needs to add color, and to put price labels on the data.

additionally fixed a bug in getOB that would cause it to overwrite the environment every time because original code used || instead of &&

updated roxygen docs and NAMESPACE